add base avatar class

+++ Test/Avatar.h
@@ -8,8 +8,8 @@ public:
 	int32 GetAvatarID() { return m_nAvatarID;}
 	void SetAvatarID(int32 nID) { m_nAvatarID = nID; }
 
-	int32 GetAvatarDID() { return m_nAvatarDID;}
-	void SetAvatarDID(int32 nDID) { m_nAvatarDID = nDID; }
+	int64 GetAvatarDID() { return m_nAvatarDID;}
+	void SetAvatarDID(int64 nDID) { m_nAvatarDID = nDID; }
 
 	bool HasAnyComState(uint64 mask) const { return (m_nComState & mask) != 0; }
 	bool HasAllComState(uint64 mask) const { return (m_nComState & mask) == mask; }
